+++ packages/app-bridge/src/react/useBlockAssets.ts
@@ -53,16 +53,11 @@ export const useBlockAssets = (appBridge: AppBridgeBlock) => {
     };
 
     const updateAssetIdsFromKey = async (key: string, newAssetIds: number[]) => {
-        const currentBlockAssets = await appBridge.getBlockAssets();
-        const oldAssetIds = currentBlockAssets[key]?.map((asset) => asset.id) ?? [];
-
         try {
-            await appBridge.deleteAssetIdsFromBlockAssetKey(key, oldAssetIds);
-            await appBridge.addAssetIdsToBlockAssetKey(key, newAssetIds);
+            await appBridge.api({ name: 'setAssetIdsByBlockAssetKey', payload: { key, assetIds: newAssetIds } });
         } catch (error) {
             console.error(error);
         }
-
         emitUpdatedBlockAssets();
     };