+++ packages/cli/templates/platform-app-css/src/index.ts
@@ -1,0 +1,10 @@
+import { App } from './App';
+import { defineApp } from '@frontify/platform-app';
+import { settings } from './settings';
+import './main.css';
+import '@frontify/fondue/style';
+
+export default defineApp({
+    app: App,
+    settings,
+});