+++ packages/guideline-blocks-settings/src/components/BlockItemWrapper/Toolbar/types.ts
@@ -1,38 +1,18 @@
 /* (c) Copyright Frontify Ltd., all rights reserved. */
 
-import { type MenuItemStyle } from '@frontify/fondue';
+import { type ToolbarButtonProps } from './ToolbarButton';
+import { type DragHandleToolbarButtonProps } from './DragHandleToolbarButton';
+import { type FlyoutToolbarButtonProps } from './FlyoutToolbarButton';
+import { type MenuToolbarButtonProps } from './MenuToolbarButton';
 
 export type ToolbarProps = {
     items: ToolbarItem[];
-    flyoutMenu: ToolbarFlyoutState & { items: FlyoutToolbarItem[][] };
-    attachments: ToolbarFlyoutState & { isEnabled: boolean };
-    isDragging?: boolean;
+    attachments: { isEnabled: boolean };
 };
 
-export type ToolbarFlyoutState = {
-    isOpen: boolean;
-    onOpenChange: (isOpen: boolean) => void;
-};
+export type DragHandleToolbarItem = { type: 'dragHandle' } & DragHandleToolbarButtonProps;
+export type ButtonToolbarItem = { type: 'button' } & ToolbarButtonProps;
+export type FlyoutToolbarItem = { type: 'flyout' } & FlyoutToolbarButtonProps;
+export type MenuToolbarItem = { type: 'menu' } & MenuToolbarButtonProps;
 
-type BaseToolbarItem = {
-    icon: JSX.Element;
-    tooltip?: string;
-};
-
-type DraghandleToolbarItem = BaseToolbarItem & {
-    draggableProps: Record<string, unknown>;
-    setActivatorNodeRef?: (node: HTMLElement | null) => void;
-};
-
-type ButtonToolbarItem = BaseToolbarItem & {
-    onClick: () => void;
-};
-
-export type ToolbarItem = DraghandleToolbarItem | ButtonToolbarItem;
-
-export type FlyoutToolbarItem = {
-    title: string;
-    onClick: () => void;
-    icon: JSX.Element;
-    style?: MenuItemStyle;
-};
+export type ToolbarItem = DragHandleToolbarItem | ButtonToolbarItem | FlyoutToolbarItem | MenuToolbarItem;